LayoutTests/fast/viewport/viewport-42.html

<head>
    <title>Viewport META with "width=device-width, height=1000, initial-scale=2".</title>
    <meta name="viewport" content="width=device-width, height=1000, initial-scale=2">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
