LayoutTests/fast/viewport/viewport-42-expected.txt

CONSOLE MESSAGE: line 3: Viewport height or height set to physical device height, try using "device-height" constant instead for future compatibility.
viewport size 320x1000 scale 2.000000 with limits [1.000000, 5.000000]
